+++ exporter/otlp-common/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### Unreleased
 
+* FEAT: Add span flags support for isRemote property in OTLP exporter [#XXXX](https://github.com/open-telemetry/opentelemetry-ruby/pull/XXXX)
 * REFACTORED: opentelemetry-exporter-otlp to extract common code here. Refer to its changelog for additional history. [#1179](https://github.com/open-telemetry/opentelemetry-ruby/pull/1179)